moved /projects page to /search/projects page
